requirements.txt
alembic==1.6.5
anyio==3.3.0
apache-airflow==2.1.3
apache-airflow-providers-ftp==2.0.0
apache-airflow-providers-http==2.0.0
apache-airflow-providers-imap==2.0.0
apache-airflow-providers-sqlite==2.0.0
apispec==3.3.2
argcomplete==1.12.3
asgiref==3.6.0
astroid==2.8.3
attrs==20.3.0
Automat==0.8.0
autopep8==1.5.7
Babel==2.9.1
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
blinker==1.4
cattrs==1.5.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
Click==7.0
clickclick==20.10.2
cloud-init==22.1
colorama==0.4.4
colorlog==6.4.1
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.7
cowsay==4.0
croniter==1.0.15
cryptography==35.0.0
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
dill==0.3.4
dinosay==1.0.0
distlib==0.3.3
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==4.1.3
django-allauth==0.46.0
django-heroku==0.0.0
django-on-heroku==1.1.2
django-reset-migrations==0.4.0
django-session-timeout==0.1.0
dnspython==2.1.0
docopt==0.6.2
docutils==0.16
durable-rules==2.0.28
email-validator==1.1.3
entrypoints==0.3
filelock==3.3.0
Flask==1.1.4
Flask-AppBuilder==3.3.2
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
fonttools==4.39.3
graphviz==0.17
greenlet==1.1.1
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.6
httplib2==0.14.0
httpx==0.19.0
hyperlink==19.0.0
idna==3.3
importlib-metadata==4.6.4
importlib-resources==5.12.0
incremental==16.10.1
inflection==0.5.1
iso8601==0.1.16
isodate==0.6.0
isort==5.9.3
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.13.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.23.1
matplotlib==3.7.1
mccabe==0.6.1
more-itertools==4.2.0
MouseInfo==0.1.3
mysql-connector-python==8.0.27
mysqlclient==2.1.0
netifaces==0.10.4
ntplib==0.4.0
numpy==1.21.2
oauthlib==3.1.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
packaging==23.1
pandas==1.3.2
passlib==1.7.4
pendulum==2.1.2
pexpect==4.6.0
Pillow==9.0.1
pipenv==2021.5.29
pipreqs==0.4.11
platformdirs==2.4.0
prison==0.1.3
protobuf==3.19.1
psutil==5.8.0
psycopg2-binary==2.9.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAutoGUI==0.9.53
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.21
pycups==1.9.73
PyGetWindow==0.0.9
Pygments==2.10.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.3.0
pylint==2.11.1
pymacaroons==0.13.0
PyMouse==1.0
PyMsgBox==1.0.9
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.1.4
pyRFC3339==1.1
pyrsistent==0.15.5
PyScreeze==0.1.28
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-daemon==2.3.0
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
python3-xlib==0.15
pytweening==1.0.4
pytz==2021.3
pytzdata==2020.1
PyYAML==5.3.1
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rfc3986==1.5.0
rich==10.7.0
scikit-learn==1.2.2
scikit-surprise==1.1.3
scipy==1.10.1
SecretStorage==2.3.1
service-identity==18.1.0
setproctitle==1.2.2
simplejson==3.16.0
six==1.16.0
sniffio==1.2.0
sos==4.3
SQLAlchemy==1.4.23
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.8
sqlparse==0.4.2
ssh-import-id==5.10
swagger-ui-bundle==0.0.8
systemd-python==234
tabulate==0.8.9
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
Twisted==18.9.0
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
unicodecsv==0.14.1
urllib3==1.26.7
virtualenv==20.8.1
virtualenv-clone==0.5.7
wadllib==1.3.3
Werkzeug==1.0.1
whitenoise==5.3.0
wrapt==1.13.2
WTForms==2.3.3
yarg==0.1.9
zipp==3.15.0
zope.interface==4.7.1
